langchain_sample/requirements.txt
langchain
langchain-google-genai
google-generativeai
python-dotenv
chromadb
tiktoken
